+++ changelogs/CHANGELOG.md
@@ -7,3 +7,7 @@
 ### ⚙️  Miscellaneous
 
 - *(init)* Project init.
+
+### 🚀 Added
+
+- *(build)* Build appears to be working now.